requirements.txt
boto3==1.18.17
botocore==1.21.17
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
eib-aws-utils==1.0.5
Flask==2.0.1
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-json-logger==2.0.2
requests==2.26.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
webexteamssdk==1.6
Werkzeug==2.0.1
